requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.23
flask-migrate==4.0.5
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0
matplotlib==3.4.3
seaborn==0.11.2
gunicorn==20.1.0
flask-wtf==1.1.1
werkzeug==2.3.7
cryptography==41.0.7
python-dotenv==1.1.1

# AI/ML Dependencies (Optional - install based on your needs)
# For TensorFlow/Keras support:
# tensorflow==2.13.0
# keras==2.13.1

# For PyTorch support:
# torch==2.0.1
# torchvision==0.15.2

# Additional ML libraries
xgboost==1.7.6
lightgbm==4.0.0